privacy-pii/requirements.txt
python-dotenv==1.0.1
regex==2025.7.34
pandas==2.3.1
numpy==2.3.2
tqdm==4.67.1

streamlit==1.48.0

transformers==4.55.0
datasets==4.0.0
tokenizers==0.21.4
huggingface-hub==0.34.4

scikit-learn==1.7.1
seqeval==1.2.2

torch==2.8.0
torchcrf==1.1.0

sklearn-crfsuite==0.5.0
python-crfsuite==0.9.11
urllib3==2.2.1
pydub==0.25.1
audioop-lts==0.2.2
imageio-ffmpeg==0.5.1